requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
certifi==2025.8.3
click==8.3.0
colorama==0.4.6
databases==0.9.0
fastapi==0.117.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.26.0
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-telegram-bot==20.8
PyYAML==6.0.2
rich==14.1.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
